Move getReactNativeVersion func to ReactNativeProjectHelper
