+++ src/common/reactNativeProjectHelper.ts
@@ -4,7 +4,6 @@
 import * as Q from "q";
 import * as fs from "fs";
 import * as path from "path";
-import {CommandExecutor} from "./commandExecutor";
 import {Package} from "./node/package";
 
 export class ReactNativeProjectHelper {
@@ -14,8 +13,17 @@ export class ReactNativeProjectHelper {
         return ["0.54.0", "0.54.1", "0.54.2", "0.54.3", "0.54.4"];
     }
 
-    public static getReactNativeVersion(projectRoot: string) {
-        return new CommandExecutor(projectRoot).getReactNativeVersion();
+    public static getReactNativeVersion(projectRoot: string): Q.Promise<string> {
+        const reactNativePackageDir = path.resolve(
+            projectRoot,
+            "node_modules",
+            "react-native"
+          );
+
+        return ReactNativeProjectHelper.getReactNativePackageVersionFromNodeModules(reactNativePackageDir)
+            .catch(err => {
+                return ReactNativeProjectHelper.getReactNativeVersionFromProjectPackage(projectRoot);
+            });
     }
 
     public static getReactNativePackageVersionFromNodeModules(reactNativePackageDir: string): Q.Promise<string> {
